backend/requirements_simple.txt
# Lucky Gas Backend - Simplified Requirements
# Only 30 essential dependencies (down from 159!)

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database - SYNC only, no async!
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pytz==2023.3

# Excel handling (for data import)
pandas==2.1.3
openpyxl==3.1.2

# Testing (optional, for development)
pytest==7.4.3
httpx==0.25.2

# Production server
gunicorn==21.2.0